backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.10.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
hnswlib==0.6.2
idna==2.10
numpy==1.22.3
psycopg2==2.9.3
py4j==0.10.9.5
pyspark==3.3.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sortedcontainers==2.2.2
soupsieve==2.3.2.post1
spark-nlp==4.0.2
sqlparse==0.4.2
tqdm==4.47.0
urllib3==1.26.10
whitenoise==6.2.0
